+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### v1.1.32
 * Actor : Added `apply_dot`
+* Survivor : Fixed some default sprites being wrong.
 
 ### v1.1.31
 * Item, Buff : Fixed lag related to internal change of `onStep` and `onDraw` callbacks.